lastBlockCheck
if running for the first time.. will index from a hard coded block value.

now blocks are maintained in a text file. when running after the second time.

fromBlock will start from getLastCheckedBlock() and check each block 1 by 1 ..

If no logs are returned from last checked Block to Current Block. nothing will be uploaded. if there are things to return. we will get an arweave link..

added a way to check 10 blocks as a fast way to check the code works.